!158 modify build.gn and seek time type for adapting to media_standard .

+++ engine/player/standard/hiplayer_impl.cpp
@@ -397,7 +397,7 @@ ErrorCode HiPlayerImpl::DoOnComplete()
     if (!singleLoop_) {
         StopAsync();
     } else {
-        fsm_.SendEventAsync(Intent::SEEK, static_cast<int64_t>(0));
+        fsm_.SendEventAsync(Intent::SEEK, static_cast<int32_t>(0));
     }
     auto ptr = obs_.lock();
     if (ptr != nullptr) {
